test(sharing): move the mock-only bulk-grant cases to the unit tier (#2000)
CONTRIBUTING.md assigns httpx_mock request-shape assertions to the Unit tier
("Reach for httpx_mock when you need to assert on outgoing request shape ...
put the test under tests/unit/"). TestAddUsers was colocated with the
sibling TestAddUser cases in the integration file instead, which is
grandfathered in integration_allow_no_vcr_files.txt as debt to migrate — so
new node IDs were quietly enlarging the exception that allowlist exists to
shrink. Both reviewers flagged it; move rather than argue.

TestAddUsers now lives in tests/unit/test_sharing_types.py, next to the
existing add_user OWNER/_REMOVE validation coverage. The request-decode
helper becomes an rpc_request_params fixture in tests/conftest.py rather
than being duplicated: both tiers assert on wire shape and
tests/_guardrails/test_no_cross_test_imports.py forbids one test module
importing another. The two assertions added to the existing integration
add_user / remove_user tests stay put (they add no node IDs) and now use
the shared fixture.

Note on the review's stated mechanism: the guardrail runs
`test_taxonomy_inventory.py --check`, which validates the taxonomy POLICY —
the script's own generated header says the node ID lists "are audit aids;
they are not a CI snapshot contract". The baseline comparison is a separate
`--check-generated` action with no call site in CI, tests, or tooling, and
it already reports both allow_no_vcr baselines stale on origin/main. So the
suite was not failing; the tier placement was still worth fixing.
