+++ tests/integration/test_sharing_integration.py
@@ -6,9 +6,6 @@ integration-tree VCR enforcement hook in ``tests/integration/conftest.py``.
 Cassette-backed coverage lives in ``tests/integration/test_vcr_comprehensive.py``.
 """
 
-import json
-from urllib.parse import parse_qs
-
 import pytest
 from pytest_httpx import HTTPXMock
 
@@ -16,12 +13,6 @@ from notebooklm import NotebookLMClient, SharePermission, ShareViewLevel
 from notebooklm.rpc import RPCMethod
 
 pytestmark = pytest.mark.allow_no_vcr
-
-
-def _request_params(request) -> list:
-    """Decode the positional params from an RPC request."""
-    outer = json.loads(parse_qs(request.content.decode())["f.req"][0])
-    return json.loads(outer[0][0][1])
 
 
 class TestGetShareStatus:
@@ -232,6 +223,7 @@ class TestAddUser:
         auth_tokens,
         httpx_mock: HTTPXMock,
         build_rpc_response,
+        rpc_request_params,
     ):
         """Test adding a user as viewer."""
         share_response = build_rpc_response(RPCMethod.SHARE_NOTEBOOK, [])
@@ -263,7 +255,7 @@ class TestAddUser:
         assert status.shared_users[1].permission == SharePermission.VIEWER
 
         requests = httpx_mock.get_requests()
-        assert _request_params(requests[0]) == [
+        assert rpc_request_params(requests[0]) == [
             [
                 [
                     "nb_123",
@@ -344,100 +336,6 @@ class TestAddUser:
         assert len(status.shared_users) == 2
 
 
-class TestAddUsers:
-    """Tests for SharingAPI.add_users()."""
-
-    @pytest.mark.asyncio
-    async def test_add_users_sends_mixed_grants_in_one_request(
-        self,
-        auth_tokens,
-        httpx_mock: HTTPXMock,
-        build_rpc_response,
-    ):
-        """A bulk grant shares per-call notification settings and refreshes once."""
-        httpx_mock.add_response(content=build_rpc_response(RPCMethod.SHARE_NOTEBOOK, []).encode())
-        httpx_mock.add_response(
-            content=build_rpc_response(
-                RPCMethod.GET_SHARE_STATUS,
-                [
-                    [
-                        ["owner@example.com", 1, [], ["Owner", "https://avatar"]],
-                        ["viewer@example.com", 3, [], ["Viewer", "https://viewer"]],
-                        ["editor@example.com", 2, [], ["Editor", "https://editor"]],
-                    ],
-                    [False],
-                    1000,
-                ],
-            ).encode()
-        )
-
-        async with NotebookLMClient(auth_tokens) as client:
-            status = await client.sharing.add_users(
-                "nb_123",
-                [
-                    ("viewer@example.com", SharePermission.VIEWER),
-                    ("editor@example.com", SharePermission.EDITOR),
-                ],
-                notify=False,
-                welcome_message="Welcome, team!",
-            )
-
-        requests = httpx_mock.get_requests()
-        assert len(requests) == 2
-        assert requests[0].url.params["rpcids"] == RPCMethod.SHARE_NOTEBOOK.value
-        assert requests[1].url.params["rpcids"] == RPCMethod.GET_SHARE_STATUS.value
-        assert _request_params(requests[0]) == [
-            [
-                [
-                    "nb_123",
-                    [
-                        ["viewer@example.com", None, SharePermission.VIEWER.value],
-                        ["editor@example.com", None, SharePermission.EDITOR.value],
-                    ],
-                    None,
-                    [0, "Welcome, team!"],
-                ]
-            ],
-            0,
-            None,
-            [2],
-        ]
-        assert [user.email for user in status.shared_users] == [
-            "owner@example.com",
-            "viewer@example.com",
-            "editor@example.com",
-        ]
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        ("grants", "message"),
-        [
-            ([], "at least one user"),
-            (
-                [("owner@example.com", SharePermission.OWNER)],
-                "Cannot assign OWNER permission",
-            ),
-            (
-                [("removed@example.com", SharePermission._REMOVE)],
-                r"Use remove_user\(\) instead",
-            ),
-        ],
-    )
-    async def test_add_users_rejects_invalid_grants(
-        self,
-        auth_tokens,
-        httpx_mock: HTTPXMock,
-        grants,
-        message,
-    ):
-        """Empty, owner, and remove grants fail before an RPC is sent."""
-        async with NotebookLMClient(auth_tokens) as client:
-            with pytest.raises(ValueError, match=message):
-                await client.sharing.add_users("nb_123", grants)
-
-        assert httpx_mock.get_requests() == []
-
-
 class TestUpdateUser:
     """Tests for SharingAPI.update_user()."""
 
@@ -484,6 +382,7 @@ class TestRemoveUser:
         auth_tokens,
         httpx_mock: HTTPXMock,
         build_rpc_response,
+        rpc_request_params,
     ):
         """Test removing a user."""
         share_response = build_rpc_response(RPCMethod.SHARE_NOTEBOOK, [])
@@ -508,7 +407,7 @@ class TestRemoveUser:
 
         requests = httpx_mock.get_requests()
         assert len(requests) == 2
-        assert _request_params(requests[0]) == [
+        assert rpc_request_params(requests[0]) == [
             [
                 [
                     "nb_123",
